refactor(services): another one bites the dust
It was the fitness module.
