+++ Cargo.toml
@@ -68,6 +68,7 @@ sea-query = "=0.31.0"
 serde = { version = "=1.0.204", features = ["derive"] }
 serde_json = "=1.0.120"
 serde_with = { version = "=3.9.0", features = ["chrono_0_4"] }
+slug = "=0.1.5"
 strum = { version = "=0.26.2", features = ["derive"] }
 # FIXME: Upgrade once https://github.com/seanmonstar/reqwest/pull/1620 is merged
 reqwest = { git = "https://github.com/thomasqueirozb/reqwest", branch = "base_url", features = [
