+++ apps/backend/Cargo.toml
@@ -19,7 +19,6 @@ chrono = { workspace = true }
 chrono-tz = { workspace = true }
 config = { path = "../../crates/config" }
 convert_case = { workspace = true }
-const-str = { workspace = true }
 csv = "=1.3.0"
 data-encoding = "=2.6.0"
 dotenvy = "=0.15.7"
@@ -50,7 +49,7 @@ serde_json = { workspace = true }
 serde_with = { workspace = true }
 serde-xml-rs = "=0.6.0"
 services = { path = "../../crates/services" }
-slug = "=0.1.5"
+slug = { workspace = true }
 tokio = { workspace = true }
 tower = { version = "=0.4.13", features = ["buffer"] }
 tower-http = { version = "=0.5.2", features = ["catch-panic", "cors", "trace"] }
