--- apps/backend/src/fitness/mod.rs
@@ -1,3 +1,0 @@
-pub mod resolver;
-
-mod logic;
