+++ crates/resolvers/fitness/Cargo.toml
@@ -4,3 +4,11 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+async-graphql = { workspace = true }
+common_models = { path = "../../models/common" }
+database_models = { path = "../../models/database" }
+dependent_models = { path = "../../models/dependent" }
+fitness_service = { path = "../../services/fitness" }
+fitness_models = { path = "../../models/fitness" }
+sea-orm = { workspace = true }
+traits = { path = "../../traits" }
